created a base.html template,
the base.html file acts as the parent of the other templates associated with it.

all the common stuff in both the other templates is written in the base, with the 'block content' in the body.

in the child templates the base template is imported and a block is created to associae with the base template.